adjustIntConst fix

+++ c2go__fpclassify.c.i.go
@@ -8,22 +8,22 @@ func __fpclassify(x float64) int32 {
 	}
 	var u _cgoa_68
 	u.f = x
-	var e int32 = int32(*(*uint64)(unsafe.Pointer(&u)) >> 52 & uint64(2047))
+	var e int32 = int32(*(*uint64)(unsafe.Pointer(&u)) >> int32(52) & uint64(2047))
 	if !(e != 0) {
 		return func() int32 {
-			if *(*uint64)(unsafe.Pointer(&u))<<1 != 0 {
-				return 3
+			if *(*uint64)(unsafe.Pointer(&u))<<int32(1) != 0 {
+				return int32(3)
 			} else {
-				return 2
+				return int32(2)
 			}
 		}()
 	}
-	if e == 2047 {
+	if e == int32(2047) {
 		return func() int32 {
-			if *(*uint64)(unsafe.Pointer(&u))<<12 != 0 {
-				return 0
+			if *(*uint64)(unsafe.Pointer(&u))<<int32(12) != 0 {
+				return int32(0)
 			} else {
-				return 1
+				return int32(1)
 			}
 		}()
 	}
