+++ getenv.c.i.go
@@ -7,7 +7,7 @@ func Getenv(name *int8) *int8 {
 	if l != 0 && !(*(*int8)(unsafe.Pointer(uintptr(unsafe.Pointer(name)) + uintptr(l))) != 0) && __environ != nil {
 		for e := (**int8)(__environ); *e != nil; *(*uintptr)(unsafe.Pointer(&e)) += 8 {
 			if !(Strncmp(name, *e, l) != 0) && int32(*(*int8)(unsafe.Pointer(uintptr(unsafe.Pointer(*e)) + uintptr(l)))) == '=' {
-				return (*int8)(unsafe.Pointer(uintptr(unsafe.Pointer((*int8)(unsafe.Pointer(uintptr(unsafe.Pointer(*e))+uintptr(l))))) + uintptr(1)))
+				return (*int8)(unsafe.Pointer(uintptr(unsafe.Pointer((*int8)(unsafe.Pointer(uintptr(unsafe.Pointer(*e))+uintptr(l))))) + uintptr(int32(1))))
 			}
 		}
 	}
