+++ test/cmd/test_snprintf/print.c.i.go
@@ -5,7 +5,7 @@ import (
 	unsafe "unsafe"
 )
 
-var t_status int32 = 0
+var t_status int32 = int32(0)
 
 func t_printf(s *int8, __cgo_args ...interface {
 }) int32 {
@@ -16,14 +16,14 @@ func t_printf(s *int8, __cgo_args ...interface {
 	t_status = int32(1)
 	ap = __cgo_args
 	n = libc.Vsnprintf((*int8)(unsafe.Pointer(&buf)), 512, s, ap)
-	if n < 0 {
+	if n < int32(0) {
 		n = int32(0)
 	} else if uint64(n) >= 512 {
 		n = int32(512)
-		*(*int8)(unsafe.Pointer(uintptr(unsafe.Pointer((*int8)(unsafe.Pointer(&buf)))) + uintptr(n-1))) = int8('\n')
-		*(*int8)(unsafe.Pointer(uintptr(unsafe.Pointer((*int8)(unsafe.Pointer(&buf)))) + uintptr(n-2))) = int8('.')
-		*(*int8)(unsafe.Pointer(uintptr(unsafe.Pointer((*int8)(unsafe.Pointer(&buf)))) + uintptr(n-3))) = int8('.')
-		*(*int8)(unsafe.Pointer(uintptr(unsafe.Pointer((*int8)(unsafe.Pointer(&buf)))) + uintptr(n-4))) = int8('.')
+		*(*int8)(unsafe.Pointer(uintptr(unsafe.Pointer((*int8)(unsafe.Pointer(&buf)))) + uintptr(n-int32(1)))) = int8('\n')
+		*(*int8)(unsafe.Pointer(uintptr(unsafe.Pointer((*int8)(unsafe.Pointer(&buf)))) + uintptr(n-int32(2)))) = int8('.')
+		*(*int8)(unsafe.Pointer(uintptr(unsafe.Pointer((*int8)(unsafe.Pointer(&buf)))) + uintptr(n-int32(3)))) = int8('.')
+		*(*int8)(unsafe.Pointer(uintptr(unsafe.Pointer((*int8)(unsafe.Pointer(&buf)))) + uintptr(n-int32(4)))) = int8('.')
 	}
-	return int32(libc.Write(1, unsafe.Pointer((*int8)(unsafe.Pointer(&buf))), uint64(n)))
+	return int32(libc.Write(int32(1), unsafe.Pointer((*int8)(unsafe.Pointer(&buf))), uint64(n)))
 }
